requirements.txt
accelerate==0.20.3
bitsandbytes==0.41.1
dadaptation==3.1
diffusers==0.18.2
embedding_reader==1.5.1
fire==0.5.0
fsspec==2023.5.0
library==0.0.0
lion_pytorch==0.0.6
matplotlib==3.7.1
multidict==6.0.4
numpy==1.22.4
Pillow==10.0.1
prodigyopt==1.0
pydantic==1.10.13
PyYAML==6.0.1
torch==2.0.1
torchvision==0.15.2
safetensors==0.3.1
scipy==1.11.3
seaborn==0.13.0
torchmetrics==1.0.3
tqdm==4.66.1
transformers==4.31.0
wordcloud==1.9.2
prettytable
wandb
xformers
clean-fid
lightning
nudenet==3.0.8
git+https://github.com/openai/CLIP.git

